+++ src/SmartObject.php
