+++ src/StaticClass.php
