+++ src/Translator.php
