+++ tests/Utils/Reflection.getReturnType.x.phpt
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * Test: Nette\Utils\Reflection::getReturnType
+ * @phpversion 7
+ */
+
+require __DIR__ . '/../bootstrap.php';
+
+class FunctionReflection
+{
+}
+
+interface FunctionReflectionFactory
+{
+
+    public function create(
+        \ReflectionFunction $reflection,
+        array $phpDocParameterTypes
+    ): FunctionReflection|string;
+
+}
+
+dump(Nette\Utils\Reflection::getReturnType(new \ReflectionMethod(FunctionReflectionFactory::class, 'create')));
